gallium: add a cap to force compute minmax indices
pipe_draw_info has min_index and max_index fields that can be useful in
indexed drawing, however gallium may decide to not compute them in some
cases to avoid impacting performance if the driver won't need them.
However, some drivers may need to always compute these values to build a
valid command stream for indexed draw.
Instead of reimplementing this functionality or having to explicitly
compute those in driver specific code, this new cap can be used to
ensure that gallium will not skip the computation.
Drivers that set this cap will be able to rely on those values to build
the command stream.

For more details on this patch:
https://lists.freedesktop.org/archives/mesa-dev/2018-March/189251.html

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>