HACK: allow sgx to create pvr_dri.so
Mesa-Loader attempts to fetch the driver to use for the backing device
using the device name reported by the kernel module. The userspace
libraries (specifically the server and server controller) for sgx rely
on the driver specifically reporting the device as 'pvr'.

SGX and ROGUE are fundamentally different and their control interfaces
are not prepared to be merged right now. Temporarily allow the sgx
gallium frontend to create the requested pvr_dri.so and associated
extensions and throw an error if people attempt to enable both at the
same time.

Signed-off-by: Randolph Sapp <rs@ti.com>