sgx_ext: roll __DRIimageExtension back from 8 to 7
The createImageFromDmaBufs extension here is implemented correctly but
DDK 1.17's implementation of buffer imports/exports was half-baked and
was never fully functional under Linux.

In addition to that, most applications expect modifiers to be supported
if DMA extensions are present. This is not true here as SGX never had
the concept of DRM modifiers and as such supports none of the required
extensions.

Signed-off-by: Randolph Sapp <rs@ti.com>